refactor: improve finalization api
